services/GoogleSheetService/requirements.txt
# Google Sheets Service Requirements
gspread>=5.10.0
google-auth>=2.20.0
google-auth-oauthlib>=1.0.0
google-auth-httplib2>=0.1.0
python-dotenv>=1.0.0